attached_assets/requirements_1766253908636.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.7
openpyxl>=3.1
python-docx>=1.1
python-multipart>=0.0.9
